requirements.txt
aiofiles==22.1.0
asyncpg==0.26.0
click==8.1.3
databases==0.6.1
greenlet==1.1.3.post0
httptools==0.5.0
importlib-metadata==5.0.0
manage.py==0.2.10
multidict==6.0.2
psycopg2-binary==2.9.4
PyYAML==6.0
sanic==22.9.0
sanic-routing==22.8.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
tabulate==0.9.0
ujson==5.5.0
uvloop==0.17.0
websockets==10.3
yoyo-migrations==8.0.0
zipp==3.9.0
